requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
pymysql==1.1.1

pydantic==2.9.2
pydantic-settings==2.5.2

python-dotenv==1.0.1

redis[async]==5.0.1

alembic==1.16.5
